requirements.txt
pyttsx3
SpeechRecognition
google-generativeai
requests
pyserial
luma.oled
python-dotenv
pandas
Pillow
